requirements.txt
google_api_python_client==2.137.0
protobuf==5.27.4
python-dotenv==1.0.1
Requests==2.32.3
